Enhancements for code maintainablitiy

+++ StudentDockerPortalProject.Grade/Models/ViewModels/GradeFormViewModel.cs
@@ -7,7 +7,7 @@ public class GradeFormViewModel
     public int Id { get; set; }
     public int StudentId { get; set; }
     public string CourseName { get; set; } = string.Empty;
-    public double Score { get; set; }
+    public decimal Score { get; set; }
     public DateTime GradeDate { get; set; } = DateTime.Today;
     public string? Notes { get; set; }
     public List<SelectListItem> StudentOptions { get; set; } = [];
